+++ manage.py
@@ -1,5 +1,10 @@
 import os
+import sys
 import unittest
+
+import xmlrunner
+
+from click import option, Choice, echo, style
 
 from app import create_test_app
 
@@ -13,6 +18,23 @@ def test():
     """Run integration tests."""
     tests = unittest.TestLoader().discover(os.path.join(os.path.dirname(__file__), "tests"))
     unittest.TextTestRunner(verbosity=2).run(tests)
+
+
+@app.cli.command()
+@option("--test-runner", type=Choice(["text", "junit"]))
+def test(test_runner: str = "text"):
+    """Run integration tests."""
+    tests = unittest.TestLoader().discover(os.path.join(os.path.dirname(__file__), "tests"))
+
+    if test_runner == "text":
+        tests_runner = unittest.TextTestRunner(verbosity=2)
+        return sys.exit(not tests_runner.run(tests).wasSuccessful())
+    elif test_runner == "junit":
+        with open("test-results.xml", "wb") as output:
+            tests_runner = xmlrunner.XMLTestRunner(output=output)
+            return sys.exit(not tests_runner.run(tests).wasSuccessful())
+
+    echo(style("Unknown Python unit test runner type", fg="red"), err=True)
 
 
 # Support PyCharm debugging