+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+* Unit test JUnit support for use in GitLab CI
+
+### Fixed
+
+* `test` CLI command returns a non-zero exit code when tests fail
 
 ### Changed